Extract soci_mutex_t into its own private header
This will allow reusing it from other SOCI files.

Also renamed scoped_lock to soci_scoped_lock for consistency.

No real changes.
